:gear: Style: API 타입 수정 및 인터페이스타입 파일 분리
