+++ src/apis/Signup/buildingSearchRequests.ts
@@ -3,6 +3,6 @@ import { ICommonResponse } from '@/types/Common/ICommonResponse';
 import { IBuilding } from '@/types/Signup/IBuilding';
 
 export const fetchBuilding = async (): Promise<ICommonResponse<IBuilding[]>> => {
-  const response = await apiClient.get('/api/room/bankList');
+  const response = await apiClient.get('/api/building/{buildingName}');
   return response.data;
 };
