+++ src/types/Common/IErrorResponse.ts
@@ -1,4 +1,3 @@
 export interface IErrorResponse {
-  res: any;
   errorMessage: string[];
 }
